MainActivity/NavigationDrawerbasic functionality
Slides in and out and is populated with the list of features. Still
needs to implement clicklisteners to change the main view when a
specific feature is selected. MainActivity is set up to handle new
fragments, etc. ALL OLD UI ELEMENTS DELETED, APP BUILDS AND RUNS
CORRECTLY. TYBG or (real this time.)